requirements.txt
# ============================================
# Mediculus Hospital - Requirements
# ============================================

# Django core
Django==5.2
djangorestframework==3.15.2

# PostgreSQL adapter
psycopg2-binary==2.9.11 --only-binary=:all:

# JWT authentication
djangorestframework-simplejwt==5.3.1

# API documentation (Swagger)
drf-spectacular==0.27.2

# Image processing (for doctor photos)
Pillow==12.1.1 --only-binary=:all:

# Environment variables
python-decouple==3.8

# CORS headers
django-cors-headers==4.4.0

# Filtering for DRF
django-filter==24.3

# WSGI server for production
gunicorn==21.2.0

# Whitenoise for static files
whitenoise==6.7.0

# Optional: for testing
pytest-django==4.8.0
